+++ mol_gnn/data/__init__.py
@@ -1,4 +1,4 @@
-from .batch import BatchMolGraph, MpnnBatch, MultiInputMpnnBatch
+from .batch import BatchedMolGraph, MpnnBatch, MultiInputMpnnBatch
 from .mixins import Datum
 from .molecule import MoleculeDatapoint, MoleculeDataset
 from .multi import MultiInputDataset
