+++ mol_gnn/utils/__init__.py
@@ -2,3 +2,4 @@ from .hparams import HasHParams, HParamsDict
 from .mixins import ReprMixin
 from .registry import ClassRegistry, Factory
 from .utils import *
+from .chem import make_mol
