+++ mol_gnn/utils/chem.py
@@ -1,0 +1,29 @@
+from rdkit import Chem
+
+
+def make_mol(smi: str, keep_h: bool, add_h: bool) -> Chem.Mol:
+    """build an RDKit molecule from a SMILES string.
+
+    Parameters
+    ----------
+    smi : str
+        a SMILES string.
+    keep_h : bool
+        whether to keep hydrogens in the input smiles. This does not add hydrogens, it only keeps them if they are specified
+    add_h : bool
+        whether to add hydrogens to the molecule
+
+    Returns
+    -------
+    Chem.Mol
+        the RDKit molecule.
+    """
+    if keep_h:
+        mol = Chem.MolFromSmiles(smi, sanitize=False)
+        Chem.SanitizeMol(
+            mol, sanitizeOps=Chem.SanitizeFlags.SANITIZE_ALL ^ Chem.SanitizeFlags.SANITIZE_ADJUSTHS
+        )
+    else:
+        mol = Chem.MolFromSmiles(smi)
+
+    return Chem.AddHs(mol) if add_h else mol
